reorganizing where some server files should be
