+++ src/client/timeline.ts
@@ -5,8 +5,8 @@ import { PARTIES } from '../server/game-data/parties';
 import { isEqual, isLower, processVersion, Version } from '../server/routes/versions';
 import { FAN_ISSUE_DATE, AS2_NEWSPAPERS, PRE_BOILER_ROOM_PAPERS, AS3_NEWSPAPERS } from '../server/game-data/newspapers';
 import { PRE_CPIP_CATALOGS, FURNITURE_CATALOGS, CPIP_CATALOGS } from '../server/game-data/catalogues';
-import { STAGE_TIMELINE } from '../server/game-logic/stage-plays';
-import { IGLOO_LISTS } from '../server/game-logic/igloo-lists';
+import { STAGE_TIMELINE } from '../server/game-data/stage-plays';
+import { IGLOO_LISTS } from '../server/game-data/igloo-lists';
 import { ROOM_MUSIC_TIMELINE, ROOM_OPENINGS, ROOM_UPDATES, TEMPORARY_ROOM_UPDATES } from '../server/game-data/room-updates';
 import { PINS } from '../server/game-data/pins';
 import {EARTHQUAKE } from '../server/game-data/updates';
