+++ src/server/routes/client-files.ts
@@ -16,13 +16,13 @@ import { MUSIC_IDS, PRE_CPIP_MUSIC_PATH } from "../game-data/music";
 import { CPIP_STATIC_FILES } from "../game-data/cpip-static";
 import { FALLBACKS } from "../game-data/fallbacks";
 import { CPIP_CATALOGS, FURNITURE_CATALOGS, IGLOO_CATALOGS, PRE_CPIP_CATALOGS } from "../game-data/catalogues";
-import { STAGE_PLAYS, STAGE_TIMELINE } from "../game-logic/stage-plays";
-import { IGLOO_LISTS } from "../game-logic/igloo-lists";
+import { STAGE_PLAYS, STAGE_TIMELINE } from "../game-data/stage-plays";
+import { IGLOO_LISTS } from "../game-data/igloo-lists";
 import { BETA_RELEASE, CAVE_OPENING_START, CPIP_UPDATE, EPF_RELEASE, MODERN_AS3, PRE_CPIP_REWRITE_DATE } from "../game-data/updates";
 import { STADIUM_UPDATES } from "../game-data/stadium-updates";
 import { As2Newspaper, AS2_NEWSPAPERS, PRE_BOILER_ROOM_PAPERS, AS3_NEWSPAPERS } from "../game-data/newspapers";
 import { CPIP_AS3_STATIC_FILES } from "../game-data/cpip-as3-static";
-import { getNewspaperName } from "../game-logic/news.txt";
+import { getNewspaperName } from "./news.txt";
 import { PINS } from "../game-data/pins";
 import { findInVersion, IdentifierMap, processTimeline, TimelineEvent, TimelineMap, VersionsTimeline } from "../game-data/changes";
 import { MIGRATOR_PERIODS } from "../game-data/migrator";
